+++ monopoly/src/main/java/ei3/monopoly/Gare.java
@@ -1,7 +1,7 @@
 package ei3.monopoly;
 
 public class Gare extends Achetable {
-    public Gare(String nom, float prix){
+    public Gare(String nom, int prix){
         super(nom,prix);
     }
 }
